ml_models/requirements.txt
fastapi
uvicorn[standard]
torch       # hoặc torch-cuXX nếu dùng CUDA
torchvision
pillow
python-multipart
numpy
